requirements.txt
Flask==1.1.1
Flask-RESTful==0.2.12
flask-swagger==0.2.14
Flask-MySQLdb==0.2.0
sentry-sdk[flask]==0.11.2
simplejson==3.8.0
pymysql==0.9.3
mysqlclient
appdirs==1.4.3
future==0.16.0
futures==3.0.5
httplib2==0.13.1
itsdangerous==1.1.0
Jinja2==2.10.1
python-dateutil==1.4.1
python-dotenv==0.10.3
PyYAML==5.1.2
six==1.12.0
virtualenv==16.6.1

